New context-sensitive search: current and .forth.
Trying a fundamental change to the text interpreter. Instead of only
searching the .forth. chain, now both the interpreter and compiler
search first "current" (the chain we are compiling into) and only if
not found there AND current not equal to .forth. do we then also
search .forth.

This should make creating assemblers and target compilers easier, since
we won't have to be quite as explicit about "quoting" - eg with \a -
words defined in current that we want to refer to (ie, compile).